+++ docs/history.rst
@@ -8,6 +8,7 @@ Version History
     ``MOUSE_RELEASED``, :ghpull:`406`.
   * bugfix: Mouse buttons 6 through 11 were erroneously decoded as left, middle, or right instead of
     ``MOUSE_BUTTON_6`` through ``MOUSE_BUTTON_11``, :ghpull:`406`.
+  * improve: performance of automatic terminal query replies, :ghpull:`405`.
   * improve: all parameterized capabilities are now memorized, about 50x faster, :ghpull:`404`.  *
     bugfix: :meth:`~Terminal.truncate` by bump of dependency ``wcwidth>=0.8.3``, :ghissue:`402`.  *
     bugfix: :meth:`~Terminal.async_inkey` dropped keystrokes while another task is busy,